30-04-2016
Commit que incluye los siguientes cambios:
-Se ha actualizado el CSS de la página, editando un poco el estilo
visual, y mejorando el aspecto de algunos elementos.
-Los elementos '<img>' ahora tienen el atributo 'alt'.
-Se han reescrito algunos de los textos en las páginas 'conocenos.php',
'servicios.php' e 'historia.php'.
-Se ha añadido una notificación a 'index.php' para informar al usuario
cuando realiza una búsqueda de qué criterios se están siguiendo según lo
que ha introducido.
-Se ha actualizado la página 'leer.php' con estos cambios: los
comentarios ahora se muestran de más antiguo a más nuevo; ahora un
usuario con permisos de administrador puede modificar/eliminar
publicaciones; se ha corregido un pequeño error que definía el botón de
'Editar' un comentario como class="boton_eliminar".
-Se ha añadido una nueva función a 'funciones.php', generarRSS, que
genera un archivo RSS en formato XML cada vez que se publica/modifica
una publicación en la base de datos.
-Se ha añadido a 'zona_lateral.php' un bloque '<div>' para que el
usuario pueda acceder a su propio perfil tras conectarse a la página con
su cuenta de Steam.
-Se ha actualizado casi por completo 'pie_pagina.php', para incluir más
información sobre la página, incluyendo enlaces a las páginas de Steam,
Twitter, Facebook, YouTube, GitHub, LaTostadora y el RSS de Rebel Gamers
Clan.

+++ acciones_post.php
@@ -72,7 +72,7 @@ if (isset($_POST['enviar_comentario']) && ($conectado_steam)) {
 	} else {
 		// Se crea la conexión a la BD para leer las publicaciones y los comentarios.
 		@$con_bd = new mysqli($datosbd['conexion'], $datosbd['usuario2'], $datosbd['contra2'], $datosbd['nombre_bd']);
-		if ($con_bd_comentarista->connect_error) {
+		if ($con_bd->connect_error) {
 			$_SESSION['cod_comentario'] = 4;
 		} else {
 			prepararConexionUTF8($con_bd);
@@ -81,38 +81,51 @@ if (isset($_POST['enviar_comentario']) && ($conectado_steam)) {
 			$id_steam = $steamprofile['steamid'];
 			$sent_prep = $con_bd->prepare("SELECT * FROM comentarios JOIN usuarios ON usuarios.id = comentarios.autor_comentario WHERE comentarios.id = (?) AND usuarios.id_steam = (?)");
 			$sent_prep_2 = $con_bd->prepare("SELECT * FROM usuarios WHERE id_steam = (?) AND permisos = (2)");
-			if (!$sent_prep || !$sent_prep_2) {
+			$progreso_correcto = false;
+			if (!$sent_prep) {
 				$_SESSION['cod_comentario'] = 5;
 			} else {
 				$sent_prep->bind_param("is", $id_comentario, $id_steam);
-				$sent_prep_2->bind_param("s", $id_steam);
-				if (!$sent_prep->execute() || !$sent_prep_2->execute()) {
+				if (!$sent_prep->execute()) {
 					$_SESSION['cod_comentario'] = 5;
 				} else {
 					$result_comentario = $sent_prep->get_result();
+					$progreso_correcto = true;
+				}
+			}
+			if (!$progreso_correcto || !$sent_prep_2) {
+				$_SESSION['cod_comentario'] = 5;
+			} else {
+				$progreso_correcto = false;
+				$sent_prep_2->bind_param("s", $id_steam);
+				if (!$sent_prep_2->execute()) {
+					$_SESSION['cod_comentario'] = 5;
+				} else {
 					$result_usuario = $sent_prep_2->get_result();
-					// Comprobamos si el es el autor del comentario o si tiene permisos de administrador.
-					if (mysqli_num_rows($result_comentario) == 1 || mysqli_num_rows($result_usuario) == 1) {
-						// Enviar comentario a la BD.
-						$sent_prep = $con_bd->prepare("UPDATE comentarios SET contenido=(?) WHERE id=(?)");
-						if (!$sent_prep) {
-							$_SESSION['cod_comentario'] = 5;
-						} else {
-							$sent_prep->bind_param("si", $texto_comentario_editado, $id_comentario);
-							if (!$sent_prep->execute()) {
-								$_SESSION['cod_comentario'] = 5;
-							} else if (mysqli_affected_rows($con_bd) == 1) {
-								$_SESSION['cod_comentario'] = 2;
-								$_SESSION['id_publicacion'] = $id_publicacion;
-								$_SESSION['id_comentario'] = $id_comentario;
-							} else {
-								$_SESSION['cod_comentario'] = 5;
-							}
-						}
+					$progreso_correcto = true;
+				}
+			}
+			
+			// Comprobamos si el es el autor del comentario o si tiene permisos de administrador.
+			if ($progreso_correcto && (mysqli_num_rows($result_comentario) == 1 || mysqli_num_rows($result_usuario) == 1)) {
+				// Enviar comentario a la BD.
+				$sent_prep = $con_bd->prepare("UPDATE comentarios SET contenido=(?) WHERE id=(?)");
+				if (!$sent_prep) {
+					$_SESSION['cod_comentario'] = 5;
+				} else {
+					$sent_prep->bind_param("si", $texto_comentario_editado, $id_comentario);
+					if (!$sent_prep->execute()) {
+						$_SESSION['cod_comentario'] = 5;
+					} else if (mysqli_affected_rows($con_bd) == 1) {
+						$_SESSION['cod_comentario'] = 2;
+						$_SESSION['id_publicacion'] = $id_publicacion;
+						$_SESSION['id_comentario'] = $id_comentario;
 					} else {
-						$_SESSION['cod_comentario'] = 8;
+						$_SESSION['cod_comentario'] = 5;
 					}
 				}
+			} else {
+				$_SESSION['cod_comentario'] = 8;
 			}
 			// Se cierra la conexión a la BD.
 			mysqli_close($con_bd);
@@ -135,35 +148,48 @@ if (isset($_POST['enviar_comentario']) && ($conectado_steam)) {
 		$id_steam = $steamprofile['steamid'];
 		$sent_prep = $con_bd->prepare("SELECT * FROM comentarios JOIN usuarios ON usuarios.id = comentarios.autor_comentario WHERE comentarios.id = (?) AND usuarios.id_steam = (?)");
 		$sent_prep_2 = $con_bd->prepare("SELECT * FROM usuarios WHERE id_steam = (?) AND permisos = (2)");
-		if (!$sent_prep || !$sent_prep_2) {
-			$_SESSION['cod_comentario'] = 5;
+		$progreso_correcto = false;
+		if (!$sent_prep) {
+			$_SESSION['cod_comentario'] = 7;
 		} else {
 			$sent_prep->bind_param("is", $id_comentario, $id_steam);
-			$sent_prep_2->bind_param("s", $id_steam);
-			if (!$sent_prep->execute() || !$sent_prep_2->execute()) {
-				$_SESSION['cod_comentario'] = 5;
+			if (!$sent_prep->execute()) {
+				$_SESSION['cod_comentario'] = 7;
 			} else {
 				$result_comentario = $sent_prep->get_result();
+				$progreso_correcto = true;
+			}
+		}
+		if (!$progreso_correcto || !$sent_prep_2) {
+			$_SESSION['cod_comentario'] = 7;
+		} else {
+			$progreso_correcto = false;
+			$sent_prep_2->bind_param("s", $id_steam);
+			if (!$sent_prep_2->execute()) {
+				$_SESSION['cod_comentario'] = 7;
+			} else {
 				$result_usuario = $sent_prep_2->get_result();
-				// Comprobamos si el es el autor del comentario o si tiene permisos de administrador.
-				if (mysqli_num_rows($result_comentario) == 1 || mysqli_num_rows($result_usuario) == 1) {
-					$sent_prep = $con_bd->prepare("DELETE FROM comentarios WHERE id = (?)");
-					if (!$sent_prep) {
-						$_SESSION['cod_comentario'] = 7;
-					} else {
-						$sent_prep->bind_param("i", $id_comentario);
-						if (!$sent_prep->execute()) {
-							$_SESSION['cod_comentario'] = 7;
-						} else if (mysqli_affected_rows($con_bd) == 1) {
-							$_SESSION['cod_comentario'] = 6;
-						} else {
-							$_SESSION['cod_comentario'] = 7;
-						}
-					}
+				$progreso_correcto = true;
+			}
+		}
+		
+		// Comprobamos si el es el autor del comentario o si tiene permisos de administrador.
+		if ($progreso_correcto && (mysqli_num_rows($result_comentario) == 1 || mysqli_num_rows($result_usuario) == 1)) {
+			$sent_prep = $con_bd->prepare("DELETE FROM comentarios WHERE id = (?)");
+			if (!$sent_prep) {
+				$_SESSION['cod_comentario'] = 7;
+			} else {
+				$sent_prep->bind_param("i", $id_comentario);
+				if (!$sent_prep->execute()) {
+					$_SESSION['cod_comentario'] = 7;
+				} else if (mysqli_affected_rows($con_bd) == 1) {
+					$_SESSION['cod_comentario'] = 6;
 				} else {
-					$_SESSION['cod_comentario'] = 8;
+					$_SESSION['cod_comentario'] = 7;
 				}
 			}
+		} else {
+			$_SESSION['cod_comentario'] = 8;
 		}
 		// Se cierra la conexión a la BD.
 		mysqli_close($con_bd);
@@ -184,35 +210,48 @@ if (isset($_POST['enviar_comentario']) && ($conectado_steam)) {
 		$id_steam = $steamprofile['steamid'];
 		$sent_prep = $con_bd->prepare("SELECT * FROM publicaciones JOIN usuarios ON usuarios.id = publicaciones.autor WHERE publicaciones.id = (?) AND usuarios.id_steam = (?)");
 		$sent_prep_2 = $con_bd->prepare("SELECT * FROM usuarios WHERE id_steam = (?) AND permisos = (2)");
-		if (!$sent_prep || !$sent_prep_2) {
+		$progreso_correcto = false;
+		if (!$sent_prep) {
 			$_SESSION['cod_publicacion'] = 4;
 		} else {
 			$sent_prep->bind_param("is", $id_publicacion, $id_steam);
-			$sent_prep_2->bind_param("s", $id_steam);
-			if (!$sent_prep->execute() || !$sent_prep_2->execute()) {
+			//$sent_prep_2->bind_param("s", $id_steam);
+			if (!$sent_prep->execute()) {
 				$_SESSION['cod_publicacion'] = 4;
 			} else {
 				$result_comentario = $sent_prep->get_result();
+				$progreso_correcto = true;
+			}
+		}
+		if (!$progreso_correcto || !$sent_prep_2) {
+			$_SESSION['cod_publicacion'] = 4;
+		} else {
+			$progreso_correcto = false;
+			$sent_prep_2->bind_param("s", $id_steam);
+			if (!$sent_prep_2->execute()) {
+				$_SESSION['cod_publicacion'] = 4;
+			} else {
 				$result_usuario = $sent_prep_2->get_result();
-				// Comprobamos si el es el autor de la publicación o si tiene permisos de administrador.
-				if (mysqli_num_rows($result_comentario) == 1 || mysqli_num_rows($result_usuario) == 1) {
-					$sent_prep = $con_bd->prepare("DELETE FROM publicaciones WHERE id = (?)");
-					if (!$sent_prep) {
-						$_SESSION['cod_publicacion'] = 4;
-					} else {
-						$sent_prep->bind_param("i", $id_publicacion);
-						if (!$sent_prep->execute()) {
-							$_SESSION['cod_publicacion'] = 4;
-						} else if (mysqli_affected_rows($con_bd) == 1) {
-							$_SESSION['cod_publicacion'] = 6;
-						} else {
-							$_SESSION['cod_publicacion'] = 4;
-						}
-					}
+				$progreso_correcto = true;
+			}
+		}
+		// Comprobamos si el es el autor de la publicación o si tiene permisos de administrador.
+		if ($progreso_correcto && (mysqli_num_rows($result_comentario) == 1 || mysqli_num_rows($result_usuario) == 1)) {
+			$sent_prep = $con_bd->prepare("DELETE FROM publicaciones WHERE id = (?)");
+			if (!$sent_prep) {
+				$_SESSION['cod_publicacion'] = 4;
+			} else {
+				$sent_prep->bind_param("i", $id_publicacion);
+				if (!$sent_prep->execute()) {
+					$_SESSION['cod_publicacion'] = 4;
+				} else if (mysqli_affected_rows($con_bd) == 1) {
+					$_SESSION['cod_publicacion'] = 6;
 				} else {
-					$_SESSION['cod_publicacion'] = 5;
+					$_SESSION['cod_publicacion'] = 4;
 				}
 			}
+		} else {
+			$_SESSION['cod_publicacion'] = 5;
 		}
 		// Se cierra la conexión a la BD.
 		mysqli_close($con_bd);
